모든 회원의 수 count long 타입으로 변환

+++ src/main/java/com/ll/exam/qsl/user/repository/UserRepositoryCustom.java
@@ -4,6 +4,6 @@ import com.ll.exam.qsl.user.entity.SiteUser;
 
 public interface UserRepositoryCustom {
     SiteUser getQslCount(Long id);
-    int getQslCount();
+    long getQslCount();
 }
